[EI-199] [web] refactor: storybook name changed & refactor
